Changes partition size of AddOnlyPersistentTrieSet from 5 to 4. Moves members of helper class PersistentTrieHelper into other classes/to package. AbstractIndexedArrayObservableSet implements now ReadOnlySequencedCollection instead of Deque.

+++ org.jhotdraw8.application/src/main/java/org.jhotdraw8.application/org/jhotdraw8/collection/IntArrayList.java
@@ -403,7 +403,7 @@ public class IntArrayList extends AbstractList<Integer> implements IntList {
 
     @Override
     public boolean add(Integer integer) {
-        addAsInt((int) integer);
+        addAsInt(integer);
         return true;
     }
